emos-image: tell Yocto that our rootfs image is read-only
This will make the build fail on failed postinst scripts rather than hiding
the error.

One consequence of this change is that dropbear keys are placed in the
volatile dir /var/lib/dropbear by default, leading to new SSH host keys
being generated after each reboot (when the first SSH session is opened).
While this SSH key is only relevant for development purposes, it is still
annoying, so we change the path to /data/var/lib/dropbear instead.

We also limit the volatile bind dirs to /var/cache and /var/lib/systemd.
In particular we avoid making the whole /var/lib writable (and volatile),
leading to a confusing volatile /var/lib/opkg, while the rest of the
filesystem is read-only.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>