conf/local.conf.sample: disable unneeded debug tweaks
The debug-tweaks image feature combines 3 different features:
empty-root-password, allow-empty-password and post-install-logging.

Disable empty-root-password, as /etc/shadow is always over-mounted with a
generated file, and we should default to a locked password rather than
having an empty password in case the mount fails for some reason.

post-install-logging is not useful either, as our root filesystem is
read-only, so no post-install scripts can ever run.

Leave in allow-empty-password, as it is useful for development.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>