nginx: reduce cache time to 1 week, do not cache certain files
We can't explicitly cache-bust index.html and apps.json. Both of these
files are tiny (< 1KB) and they are never reloaded while the user navigates
the EM web interface, so we just prevent caching them altogether.

As a nice side effect, it is not possible to open a non-functional cached
web interface for an unreachable EM anymore: the user will get an error
messages from the browser right away.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>