YODA-516: micrel-netdev-daemon: set eth0 down when no switch port has a link
The micrel-netdev-led-daemon is renamed to micrel-netdev-daemon, as it
only handles the link status now. The netdev-led package now handles the
LED on EM310 hardware as well.

The eth0 is set down when no link is detected on either switch port. As
soon as there is a link, systemd-networkd is restarted to trigger a new
DHCP request.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>