conf/distro/emos.conf: use libmodbus 3.1.x
libmodbus 3.0.6 can hang in connect() for very long times (> 2 minutes)
when no connection is possible, making the modbus-daemon hang when trying
to stop it. libmodbus 3.1.x fixes this issue.

Note that libmodbus 3.1.x is incompatible with libmodbus 3.0.x, leading
to segfaults when a modbus-daemon built against 3.0.x is run with 3.1.x.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>