em-image-core: add minimal image
This minimal image does not contain any apps (and should not be installed
on an EM, as it is missing even micrel-switch-tool), but it can be built
only using openembedded-core, meta-openbedded, meta-rauc and meta-emos.

It can be used to create an EM-compatible toolchain package, or a base
image that is combined with prebuilt app packages to create a full-featured
firmware.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>