emit: add "Energy Manager Image Tool"
The image tool takes a core image and a list of binary app packages
(downloaded via HTTPS and creates a signed RAUC bundle from this).

The bundle contents are defined by a YAML spec file. There are two such
spec files:
- The base spec, containing core apps that are included in most images
- The bundle spec, defining additional apps and product info strings

The default base spec is part of the emit package itself; it usually
doesn't need to be replaced. The bundle spec is defined by the bundle
recipe itself.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>