empkg: update D-Bus policies earlier, clean up obsolete files
We need to update the policies before we try to (re)start the systemd
units, or the start may fail if the policy has changed.

In addition, we make sure that old policy files for apps that are still
installed and enabled, but don't have a policy file anymore are cleaned up
correctly.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>