empkg: allow starting and stopping all non-core apps using em-app.target
In addition to moving the non-core apps from multi-user.target.wants to
em-app.target.wants, a PartOf setting is added, so these apps are stopped
together with em-app.target.

For now, we hardcode the list of core apps. In a future version, this might
use the "essential" flag from the app manifest instead, or another separate
manifest flag.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>