emcfg: store Teridian meter registers on shutdown
The Teridian firmware stores its registers only once an hour. As the
Teridian chip is reset on reboots, this leads to the loss of up to one hour
of measurement data.

This becomes especially apparent when a backup restore was performed before
rebooting: the restored register values are lost. Fix this by always
storing the register data on shutdown.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>