emos-upgrade: clear inactive partition after reboot
emos-upgrade-finalize is run after rauc-mark-good - that means that the
current boot is considered successful. In this state, there is no sane
way to return to the other root filesystem: In the previous setup, we
kept both filesystems identical, but we had no way to determine that
they had become non-identical, so the second image was never used again
anyways.

Get rid of some complexity and improve boot time by removing the second
upgrade completely. Instead, we overwrite the inactive partitions with
zeros after reboot - we don't even need to care whether this inactive
partition is the old rootfs after a successful upgrade, or the new
rootfs that failed to boot 3 times.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>