emit: install images regardless of stored checksums
Clearing the inactive partition manually modifies its contents without
RAUC noticing, so it will skip the actual image write when upgrading
with the same image that was installed before. In consequence, a boot
slot which isn't actually bootable becomes active.

Disable the "intelligent" updater logic which skips writes. This has
become the default in newer RAUC versions, but we are still on RAUC 1.1,
so we need to set this manually.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>