recipes-core/emcfg/files/em-init: add EOL status handling
The health-check app checks the EOL status register of the eMMC once a day.
If the status is urgent, it sets the EOL flag in the u-boot environment and
executes a reboot.
The init process provides an EOL mode with reduced functionality. In this mode
the flash is written little and the customer still can save his data.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>