recipes-core/emcfg/files/em-init: add file system checks
The init script checks relevant partitions with tune2fs
if there is a valid maximum mount count. If not it sets a default values
for maximum mount count and check interval.

The file system checks are triggered every reboot but will only doing the
check if a file system seems not clean. This is the case with an expired
maximum mount count or when the check interval has been exceeded.

In the event of errors, the status and error log are saved in /run and
can later be used by the application.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>