recipes-core/emcfg: remove em-store-teridian-registers
The meter readings are saved every hour in the Teridian chip. To prevent data loss when restarting,
the registers were previously stored in the service 'em-store-teridian-registers.service' when the
device was shut down.

In the backup export, only the teridiand app is briefly closed and then restarted,
em-store-teridian-registers.service is not executed. When the teridiand app is started, the
Teridian chip is restarted, which loses meter readings that have not been saved. As a result,
the values of the meter readings spring back.

The problem is solved by deleting the 'em-store-teridian-registers.service' service and saving
the registers when the teridiand app is closed.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>