em-network-config: add switch-based network configuration
A bridge br0 is created, connecting the two switch ports lan0 and lan1.
The configuration that was previously applied to eth0 is moved to br0
instead.

To avoid differences between EM300 and EM310 configuration, the bridge
is also created on EM300 (where eth0 is its only port). The existence of
a switch is detected at runtime, so the same image can be used on both
EM4xx-CB variants.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>