em-power-handler: add package
em-power-handler replaces the manual reset through the device settings
app in case of power faults (overcurrent) of the RS485 supply with a
separate script. Instead of configuring the supply enable/fault pins as
an LED and a button, they are handling using the libgpiod tools.

The new script changes the handling so that:

- The enable pin is deasserted as soon as a fault is detected. This is
  required for the new hardware revision, which does not implement this
  behaviour in hardware, but it is also compatible with the old hardware
  revision
- The power is enabled automatically every 1s. This avoids the need to
  reset the power manually, which would be inconvenient for the USB port
  of the new hardware revision

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>