empkg: add user/group handling
This enables to start a service as a non-root user and/or group.

A new non-root user is created during boot by 'empkg sync' or
'empkg install <app>', if a user exists in systemd service file.
If user or group of an app is changed, old user/group entries are removed.
The ownership of the working directory and the runtime directory is handled.

A 'Group' entry in the systemd service file can optionally be added.
This enables more than one service to use a common group which may already exist.

These systemd service file entries lead to the following ownerships,
while '-' means that no entry exists:

 User    | Group  | Ownership         |
---------|--------|-------------------|
-        | -      | root:root         |
service1 | -      | service1:service1 |
service1 | group1 | service1:group1   |
-        | group1 | root:group1       |
service1 | root   | service1:root     |

Furthermore the creation of the runtime directory is handled in empkg,
so that app do not need to create it in ExecStartPre.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>