emit: improve cache handling
The bundle definitions need to have a cache setting for every app so far.
In most cases all cache settings are the same, so we a lot of redundant information.

On the one hand these changes allow to have no cache setting at all,
this defaults to 'cache = False'. On the other hand it allows to set
a cache setting in defaults, which is the new default for the bundle,
but always can be overridden from the app entry.

The new order of precedence for the cache settings are:
1. value in app
2. value in defaults
3. No value, which defaults to False

Signed-off-by: Christoph Krutz <Christoph.Krutz@tq-group.com>