systemd: increase priority of systemd and decrease priority of long tasks
Systemd daemon-reload may take a long time due to other process taking
to much resources. Systemd taking too long may result in an unintended
watchdog timeout an hardware reboot. In order to reduce the risk of
this, we increase the priority of systemd and decrease the priority of
less important processes to give systemd a larger share of processing
time and finish faster.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>