systemd: increase wdt timeout to a safe value
The watchdog timed out in a version with a wdt of 29s
because of a long systemd daemon reload of 30s on hw0100.
To be on the safe side we calculate with a 30s long
systemd daemon reload and a refresh rate of 30s and leaving
a safety margin of 15s leading to 75s of watchdog timeout.

Signed-off-by: Stöter Thomas <thomas.stoeter@tq-group.com>