journald: keep single system-journal
A side effect of the current initiative to
isolate apps from each other is that journald
in its default configuration creates separate
journals per each user whose id is above 1000
(which is the case for the isolated app users).

Not only is this not desired, it also rapidly
consumes free space on the partition containing
the journal log.

This commit changes the default option for
journald to keep a single journal file nevertheless.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>