volatile-binds: generalize bindings
Since Yocto Kirkstone the ConditionPathIsReadWrite 
in systemd changed to require an existing path to be true.
(That is the parent of the first directory)

The path /var/volatile/lib does not exist
because it lies on a newly created tmpfs.

This results in the generated services mounting
/var/volatile/lib/[A,B,C] to /var/lib/[A,B,C]
not being executed and depending services
not finding a writable /var/lib/[A,B,C].

Change the service directory pattern one level up
to make the condition true.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>