machine: list bootloader filenames in machine configurations
To avoid hardcoding more machine-specific configuration in emit, add
information on bootloader filenames to the machine configs.

This also prepares for supporting multiple different bootloader images
in a single machine. Instead of specifying only a single image, a list
compatibles is mapped to filenames. The compatible string is the first
element of the toplevel Device Tree compatible. In the Bitbake varflags,
the comma is replaced with an underscore, as it is unclear if including a
comma is well-supported.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>