u-boot-imx-em: include trusted-firmware-a and firmware-imx-8m license information
As trusted-firmware-a and firmware-imx-8m only appear as build
dependencies of U-Boot and not as runtime dependencies, their license
information was never installed to the core image, even though they are
included in the final U-Boot binary.

Introduce a bbclass extra-license-depends that can be used to import all
license information from one package into another for the purpose of
deloyment to LICENSE_DIRECTORY (tmp/deploy/licenses).

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>