k3-bootloader-image: new recipe
The k3-bootloader-image takes every variant of the R5 SPL and combines
it with the A53 SPL and U-Boot proper to a single image that can be
flashed to the eMMC boot partitions.

The recipe assumes to be called from a multiconfig that has the same
name as the machine for the A53 bootloader build. The R5 SPL must be
built by a multiconfig with the same name suffixed with "-k3r5". This
differs from the setup found in regular meta-ti builds, where the A53
bootloader is built by the main/default multiconfig, and the R5
multiconfig is just called "k3r5".

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>