machine: add em-aarch64
This creates the unified em-aarch64, which references the em4xx and
em-cb30 machines as subconfigs and builds a bundle which contains
bootloader images for both platforms. The system can detect the hardware
variant at runtime, in order to:

- select the correct U-Boot environment offset and default contents
- select the correct bootloader image during upgrades
- display the correct hardware name in the Device Settings and provide a
  hardware-specific product ID via Modbus

The same kernel image is used for all platforms, and Device Trees for
both EM4xx and EM-CB30 are included.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>