trusted-firmware-a: move settings to machine configs, remove SRCREV override
We can just use whatever version is provided by meta-arm - the override
was only necessary when meta-arm's version was too old. We can also get
rid of the bbappend by moving the configuration to our machine configs.

At the moment this results in a configuration and build that is entirely
identical to the old one, as our SRCREV override was the same that
meta-arm contains.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>