em-aarch64: make em4xx and em-cb30 support optional
Building em-cb30 support when it is not actually needed can take
significant time and disk space due to the second (ARMv7) toolchain
required for the R5 bootloader.

Add options EM_AARCH64_em4xx and EM_AARCH64_em-cb30 that can be set to "0"
in local.conf to disable including these multiconfigs. This affects not
only the bootloader list included in the bundle, but also the core image,
which will be missing fw_env.config and u-boot-initial-env for the
disabled multiconfigs.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>