emos: add seccomp to allow systemd seccomp features
Add seccomp to the DISTRO_FEATURES to allow more sandboxing features of
systemd. This adds libseccomp to the distro and allows seccomp
sandboxing features like systemcall filtering.

Use DISTRO_FEATURES instead of DISTRO_FEATURES_DEFAULT as this is the
variable that is meant to be used to customize distribution features.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>