em-power-handler: solve race condition
Calling gpiomon and gpioset immediately after one another
as background processes could result in a race between
the two.

In the error case, once the POWER_GPIO enables power on the
RS485 bus, the FAULT_GPIO (low active) is set low by an
external power monitor.

The transition to low could be missed by gpiomon
if gpioset happens to win the race.

This results in em-power-handler being stuck in
waiting for gpiomon to exit.

Fix this by postponing the start of gpioset.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>