empkg: add deferred state
When empkg sync is called from em-init all apps are synced
alphabetically. Therefore it is possible that paths owned by other apps
are not yet available. This would cause an error.

Add an ERRORDEFER return to empkg_users_handle_permissions to allow the
calling function in empkg_sync to initialise the deferred apps again.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>