bundles: update em-bundle-sdk and split apps into yml files
To allow the reduction of apps in the sdk bundle and creation of a new
bundle remove the apps-block from base.yml and add it to mandatory,
development, integrable and core-qa yml files.

Change sdk bundle to use only mandatory and development apps.

Add emcore-qa bundle with all apps.

While at it update all apps to latest version.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>
Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>