classes: extra-license-depends fix directory paths
Since commit a31b46627a85 ("license: Fix directory layout issues") in
openembedded-core the license directories now contain the architecture.

Updated the paths for license dependencies to include
LICENSE_DEPLOY_PATHCOMPONENT. This ensures that the correct
directories are used for license deployment and as source.

For EXTRA_LICENSE_MCDEPENDS the toconfig has to be included.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>