emit: fail bundle build if rootfs too large for device
The effective default rootfs size was determined through
a simple write test to an empty filesystem. That size
of course depends on what the rootfs partition size is
and may change in the future.

In case that hard limit is not valid anymore
the bundle spec can define a custom 'rootfs_limit'
in its 'defaults' section.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>