udevd: ttyAPP: introduce priority
When the device files for ttymxc1/2 and ttyS1/2 exist at the same time
udev assigns the symlink "ttyAPP1/2" to whichever is created first.
This results in a race condition for the target of that symlink.
In cases where ttyAPP1/2 point to ttyS1/2, respectively, on a
device that has ttymxc1/2 as its serial interface (RS485),
there is no serial communication over the symlinks possible.

Introduce a priority flag to make udevd use the correct
link target.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>