em-appctl: fix reset of disabled apps
The reset command should remove config files
without the app running. It asks systemd
to stop the corresponding services in advance.

If an app is enabled, its service is registered
with systemd and can be started and stopped.

Disabled apps are not registered, hence systemd
does not know these services and returns an error.

em-appctl should therefor not ask systemd
about disabled apps.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>