empkg: add support for minimal system memory
Introduce a new flag 'minsysmem' that is determined
by the field 'min_sys_mem' in the app manifest.
It allows to define the amount of system memory
(in bytes or human-readable format) required as minimum for the app.

While at it, fix a compiler warning about
disregarding a const qualifier and move the
message about maximum deferral count to the correct
position.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>