empkg: link WantedBy dependencies
If service "X" has a reverse dependency like:

X.service:
  [Install]
  WantedBy=Y.target

then systemd evaluates it as forward dependency:

Y.target:
  [Unit]
  Wants=X.service

In our case the Y.target is installed on the rootfs (read-only!)
and additional dependencies upon that target are dynamically
added through the service files.

The em-app-generator must declare the forward dependency
at runtime. It scans each app's service file for a
'WantedBy' key and applies the relation to the
first unit defined as target.

Signed-off-by: Michael Krummsdorf <michael.krummsdorf@tq-group.com>